Add sepreate tests for process_job function
